+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Improved
+- Added a option to choose how much orders show in HRX Orders list
+
 ## [1.0.1] - 2022-12-09
 ### Improved
 - Added marker logo display by country