+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Improved
 - Added a option to choose how much orders show in HRX Orders list
+- Disabled mass action checkbox for orders, which have error
 
 ## [1.0.1] - 2022-12-09
 ### Improved